CLI: initialize state and profiles - create state service (readonly)  in node - split state service between node and main - split profiles service between node and main - set up and init state and profiles services in cli
