+++ src/vs/platform/state/node/state.ts
@@ -4,18 +4,14 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
-import { IStateService } from 'vs/platform/state/node/state';
 
-export const IStateMainService = createDecorator<IStateMainService>('stateMainService');
+export const IStateService = createDecorator<IStateService>('stateService');
 
-export interface IStateMainService extends IStateService {
+export interface IStateService {
 
 	readonly _serviceBrand: undefined;
 
-	setItem(key: string, data?: object | string | number | boolean | undefined | null): void;
-	setItems(items: readonly { key: string; data?: object | string | number | boolean | undefined | null }[]): void;
+	getItem<T>(key: string, defaultValue: T): T;
+	getItem<T>(key: string, defaultValue?: T): T | undefined;
 
-	removeItem(key: string): void;
-
-	close(): Promise<void>;
 }
